Moved out some report types to fix import bug
